Float support added, temporary mechanism for measurmenst using GPIO added

+++ Inc/prox_sensor.h
@@ -67,9 +67,9 @@ typedef struct
 	uint8_t BwTh_B;
 
 	/* Threshold for image binarization */
-	uint8_t Grayscale_coeff_R;
-	uint8_t Grayscale_coeff_G;
-	uint8_t Grayscale_coeff_B;
+	float Grayscale_coeff_R;
+	float Grayscale_coeff_G;
+	float Grayscale_coeff_B;
 }ProxSensor_Config_T;
